flux: set FLUX_SHELL_PLUGIN_NAME for mpibind plugin
Future versions of flux-core will require FLUX_SHELL_PLUGIN_NAME
to be defined before including flux/shell.h. This macro will be used
to define the component name for logging messages, so if it is missing
a compilation error will result.

Since it does no harm to define it when unneeded, set it to "mpibind"
at the top of the Flux shell plugin definition to avoid future compilation
errors when building against newer flux-core installations.