flux: load hwloc XML from shell when MPIBIND_TOPOFILE not set
Problem: When MPIBIND_TOPOFILE is not set, the mpibind plugin loads the
topology from the local system with hwloc_topology_load(). However,
this is known to be a very heavyweight operation, and the hwloc XML
is available from the job shell, so the mpibind plugin might as well
use that.

If MPIBIND_TOPOFILE is not set, use flux_shell_get_hwloc_xml(3)
to fetch the job shell's copy of the XML and avoid rescanning the
entire topology.