flux: fix mpibind plugin in nested Flux jobs (#31)
* flux: fix whitespace in plugin.c

Problem: Many lines in flux/plugin.c end in whitespace, which is
generally discouraged.

Fix up whitespace in all lines of flux/plugin.c.

* flux: load hwloc XML from shell when MPIBIND_TOPOFILE not set

Problem: When MPIBIND_TOPOFILE is not set, the mpibind plugin loads the
topology from the local system with hwloc_topology_load(). However,
this is known to be a very heavyweight operation, and the hwloc XML
is available from the job shell, so the mpibind plugin might as well
use that.

If MPIBIND_TOPOFILE is not set, use flux_shell_get_hwloc_xml(3)
to fetch the job shell's copy of the XML and avoid rescanning the
entire topology.

* flux: restrict topology to current cpuset if necessary

Problem: When mpibind is active in a nested Flux job, it can assign
the incorrect cores because Flux "reranks" assigned cores to zero
origin, but mpibind assigns cpusets based on the whole system topology.
For example, a child instance may be assigned OS cores 3-4, but these
appear as cores 0-1 within the instance. The mpibind plugin may then
assign OS core 0 or 1 to a job within this instance, instead of OS
cores 3 or 4.

Restrict the hwloc topology to the current cpuset mask before assigning
cores in the mpibind pluign. This fixes the issue since Flux will have
restricted its cpuset to only the OS cores to which it has access,
and hwloc will then use the logical cores reported by the job shell
as an index into these accessible resources.

There is no need to restrict the hwloc topology if the XML was obtained
from the job shell, since the Flux instance has already taken care
of restricting the provided XML.