flux: prefer loading hwloc topology from job shell by default
Problem: It is advantageous to load the hwloc topology XML from the
job shell when mpibind is running under Flux, not only because this
is likely slightly faster than loading a file from the filesystem,
but also because Flux has already restricted the hwloc topology to
only those resources available to the current job. However, the Flux
plugin currently prefers to load from a file if MPIBIND_TOPOFILE
is set, which creates a quandary for system administrators in mixed
environments, where MPIBIND_TOPOFILE should continue to be set for use
in Slurm and other execution environments, but unset for use with Flux.

To resolve, have the mpibind Flux plugin ignore MPIBIND_TOPOFILE by
default, unless FLUX_MPIBIND_USE_TOPOFILE is set in the environment.
Thus, the Flux plugin will use the job shell provided hwloc XML unless
explicitly overidden, even if MPIBIND_TOPOFILE is set, for example
on mixed clusters or clusters running only Slurm.